Adjust the lighting in the scene
With your structure built, you might want to adjust the lighting. In a new 3D scene, an object named Directional Light is included to simulate sunlight. You can manipulate the direction of this light by rotating it. Changing the position or scale of the Directional Light GameObject has no effect.