Add a Physic material
Earlier, you added a Rigidbody component to a GameObject to give it physical properties. You can also add additional physical characteristics with another type of material called Physic materials. A Physic material is a different type of material that makes an object bounce and changes its friction and drag properties.These properties take effect when the object is under the effects of gravity. Physic materials don’t change the visual properties of the GameObject.