Import assets from the Asset Store
Select the Open in Unity button for the assets you just selected. If your computer prompts you to open the Unity Editor, select the option to do so. The Unity Editor will open with the My Assets section of the Package Manager displayed, including the asset package you just selected in the Asset Store.